pos-processor/requirements.txt
Flask
gunicorn
google-cloud-bigquery
jsonschema